backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
python-multipart==0.0.20
webdavclient3==3.14.6
cryptography==44.0.0
pydantic==2.10.4
pywinpty==2.0.14
